# web/openapi.json
{
  "components": {
    "schemas": {
...
      "CommonResponse": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse",
        "type": "object"
      },
...
      "CommonResponse_Any_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[Any]",
        "type": "object"
      },
...
      "CommonResponse_DeleteManagedHostResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeleteManagedHostResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[DeleteManagedHostResponse]",
        "type": "object"
      },
...
      "CommonResponse_ListManagedHostImagesResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ListManagedHostImagesResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[ListManagedHostImagesResponse]",
        "type": "object"
      },
...
      "CommonResponse_ManagedHostSchema_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ManagedHostSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[ManagedHostSchema]",
        "type": "object"
      },
      "CommonResponse_PullManagedHostImagesResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PullManagedHostImagesResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[PullManagedHostImagesResponse]",
        "type": "object"
      },
...
      "CommonResponse_QueryManagedHostsResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QueryManagedHostsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[QueryManagedHostsResponse]",
        "type": "object"
      },
...
      "CreateManagedHostRequest": {
        "properties": {
          "docker_client_ca_cert": {
            "default": "",
            "title": "Docker Client Ca Cert",
            "type": "string"
          },
          "docker_client_cert": {
            "default": "",
            "title": "Docker Client Cert",
            "type": "string"
          },
          "docker_client_key": {
            "default": "",
            "title": "Docker Client Key",
            "type": "string"
          },
          "docker_management_port": {
            "default": 2375,
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Docker Management Port",
            "type": "integer"
          },
          "docker_tls_enabled": {
            "default": false,
            "title": "Docker Tls Enabled",
            "type": "boolean"
          },
          "host_account": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Host Account",
            "type": "string"
          },
          "host_password": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Host Password",
            "type": "string"
          },
          "ip_address": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Ip Address",
            "type": "string"
          },
          "ssh_port": {
            "default": 22,
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Ssh Port",
            "type": "integer"
          }
        },
        "required": [
          "ip_address",
          "host_account",
          "host_password"
        ],
        "title": "CreateManagedHostRequest",
        "type": "object"
      },
...
      "DeleteManagedHostImageRequest": {
        "properties": {
          "force": {
            "default": false,
            "title": "Force",
            "type": "boolean"
          },
          "image_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Image Id",
            "type": "string"
          }
        },
        "required": [
          "image_id"
        ],
        "title": "DeleteManagedHostImageRequest",
        "type": "object"
      },
      "DeleteManagedHostResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id"
        ],
        "title": "DeleteManagedHostResponse",
        "type": "object"
      },
...
      "ListManagedHostImagesResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ManagedHostImageSchema"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ListManagedHostImagesResponse",
        "type": "object"
      },
...
      "ManagedHostImageSchema": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "image_hash": {
            "default": "",
            "title": "Image Hash",
            "type": "string"
          },
          "image_id": {
            "default": "",
            "title": "Image Id",
            "type": "string"
          },
          "image_name": {
            "title": "Image Name",
            "type": "string"
          },
          "image_size": {
            "default": 0,
            "format": "int64",
            "title": "Image Size",
            "type": "integer"
          }
        },
        "required": [
          "image_name"
        ],
        "title": "ManagedHostImageSchema",
        "type": "object"
      },
      "ManagedHostSchema": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "docker_client_ca_cert": {
            "title": "Docker Client Ca Cert",
            "type": "string"
          },
          "docker_client_cert": {
            "title": "Docker Client Cert",
            "type": "string"
          },
          "docker_client_key": {
            "title": "Docker Client Key",
            "type": "string"
          },
          "docker_management_port": {
            "title": "Docker Management Port",
            "type": "integer"
          },
          "docker_tls_enabled": {
            "title": "Docker Tls Enabled",
            "type": "boolean"
          },
          "host_account": {
            "title": "Host Account",
            "type": "string"
          },
          "host_password": {
            "title": "Host Password",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "ip_address": {
            "title": "Ip Address",
            "type": "string"
          },
          "ssh_port": {
            "title": "Ssh Port",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ip_address",
          "ssh_port",
          "host_account",
          "host_password",
          "docker_management_port",
          "docker_tls_enabled",
          "docker_client_ca_cert",
          "docker_client_cert",
          "docker_client_key",
          "created_at",
          "updated_at"
        ],
        "title": "ManagedHostSchema",
        "type": "object"
      },
      "PullManagedHostImageResultSchema": {
        "properties": {
          "image_name": {
            "title": "Image Name",
            "type": "string"
          },
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "image_name",
          "success"
        ],
        "title": "PullManagedHostImageResultSchema",
        "type": "object"
      },
      "PullManagedHostImagesRequest": {
        "properties": {
          "image_names": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Image Names",
            "type": "array"
          }
        },
        "required": [
          "image_names"
        ],
        "title": "PullManagedHostImagesRequest",
        "type": "object"
      },
      "PullManagedHostImagesResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PullManagedHostImageResultSchema"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PullManagedHostImagesResponse",
        "type": "object"
      },
...
      "QueryManagedHostsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ManagedHostSchema"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "size",
          "total",
          "items"
        ],
        "title": "QueryManagedHostsResponse",
        "type": "object"
      },
...
      "UpdateManagedHostRequest": {
        "properties": {
          "docker_client_ca_cert": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Docker Client Ca Cert"
          },
          "docker_client_cert": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Docker Client Cert"
          },
          "docker_client_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Docker Client Key"
          },
          "docker_management_port": {
            "anyOf": [
              {
                "maximum": 65535.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Docker Management Port"
          },
          "docker_tls_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Docker Tls Enabled"
          },
          "host_account": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host Account"
          },
          "host_password": {
            "anyOf": [
              {
                "maxLength": 512,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host Password"
          },
          "ip_address": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "ssh_port": {
            "anyOf": [
              {
                "maximum": 65535.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssh Port"
          }
        },
        "title": "UpdateManagedHostRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AccessTokenAuth": {
        "in": "header",
        "name": "X-Z3r0-Access-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Z3r0 - A controlled multi-agent workbench for authorized security assessment, code auditing, internal review, and controlled research.",
    "version": "0.2.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/hosts": {
      "get": {
        "operationId": "query_managed_hosts_route_api_hosts_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "title": "Keyword",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_QueryManagedHostsResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Query Managed Hosts Route",
        "tags": [
          "hosts"
        ]
      },
      "post": {
        "operationId": "create_managed_host_handler_api_hosts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManagedHostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_ManagedHostSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Create Managed Host Handler",
        "tags": [
          "hosts"
        ]
      }
    },
    "/api/hosts/{id}": {
      "delete": {
        "operationId": "delete_managed_host_handler_api_hosts__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_DeleteManagedHostResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Managed host not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Delete Managed Host Handler",
        "tags": [
          "hosts"
        ]
      },
      "patch": {
        "operationId": "update_managed_host_handler_api_hosts__id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateManagedHostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_ManagedHostSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Managed host not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Update Managed Host Handler",
        "tags": [
          "hosts"
        ]
      }
    },
    "/api/hosts/{id}/images": {
      "get": {
        "operationId": "list_managed_host_images_handler_api_hosts__id__images_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_ListManagedHostImagesResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Managed host not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "List Managed Host Images Handler",
        "tags": [
          "hosts"
        ]
      }
    },
    "/api/hosts/{id}/images/pull": {
      "post": {
        "operationId": "pull_managed_host_images_handler_api_hosts__id__images_pull_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PullManagedHostImagesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_PullManagedHostImagesResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Managed host not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Pull Managed Host Images Handler",
        "tags": [
          "hosts"
        ]
      }
    },
    "/api/hosts/{id}/images/remove": {
      "post": {
        "operationId": "delete_managed_host_image_handler_api_hosts__id__images_remove_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteManagedHostImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Managed host not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Delete Managed Host Image Handler",
        "tags": [
          "hosts"
        ]
      }
    },
...
  }
}